+++ gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/main/java/io/gravitee/rest/api/portal/rest/resource/auth/OAuth2AuthenticationResource.java
@@ -22,11 +22,7 @@ import java.io.IOException;
 import java.security.KeyManagementException;
 import java.security.NoSuchAlgorithmException;
 import java.util.Base64;
-import java.util.Collection;
-import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
 import java.util.stream.Collectors;
 
@@ -37,7 +33,6 @@ import javax.net.ssl.TrustManager;
 import javax.servlet.http.HttpServletResponse;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
-import javax.ws.rs.BadRequestException;
 import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -59,33 +54,18 @@ import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.context.SecurityContextHolder;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.jayway.jsonpath.JsonPath;
-import com.jayway.jsonpath.ReadContext;
 
 import io.gravitee.common.http.MediaType;
-import io.gravitee.el.TemplateEngine;
 import io.gravitee.el.spel.function.JsonPathFunction;
-import io.gravitee.repository.management.model.MembershipDefaultReferenceId;
 import io.gravitee.repository.management.model.MembershipReferenceType;
 import io.gravitee.repository.management.model.RoleScope;
 import io.gravitee.rest.api.idp.api.authentication.UserDetails;
-import io.gravitee.rest.api.model.GroupEntity;
-import io.gravitee.rest.api.model.NewExternalUserEntity;
 import io.gravitee.rest.api.model.RoleEntity;
-import io.gravitee.rest.api.model.UpdateUserEntity;
 import io.gravitee.rest.api.model.UserEntity;
-import io.gravitee.rest.api.model.configuration.identity.GroupMappingEntity;
-import io.gravitee.rest.api.model.configuration.identity.RoleMappingEntity;
 import io.gravitee.rest.api.model.configuration.identity.SocialIdentityProviderEntity;
 import io.gravitee.rest.api.portal.rest.model.PayloadInput;
 import io.gravitee.rest.api.portal.rest.utils.BlindTrustManager;
-import io.gravitee.rest.api.service.GroupService;
-import io.gravitee.rest.api.service.MembershipService;
-import io.gravitee.rest.api.service.RoleService;
 import io.gravitee.rest.api.service.SocialIdentityProviderService;
-import io.gravitee.rest.api.service.exceptions.GroupNotFoundException;
-import io.gravitee.rest.api.service.exceptions.RoleNotFoundException;
-import io.gravitee.rest.api.service.exceptions.UserNotFoundException;
 import io.swagger.annotations.Api;
 
 /**
@@ -99,28 +79,10 @@ public class OAuth2AuthenticationResource extends AbstractAuthenticationResource
 
     private static final Logger LOGGER = LoggerFactory.getLogger(OAuth2AuthenticationResource.class);
 
-    private static final String TEMPLATE_ENGINE_PROFILE_ATTRIBUTE = "profile";
-
     @Autowired
     private SocialIdentityProviderService socialIdentityProviderService;
 
-    @Autowired
-    private GroupService groupService;
-
-    @Autowired
-    private RoleService roleService;
-
     private Client client;
-
-    // Dirty hack: only used to force class loading
-    static {
-        try {
-            LOGGER.trace("Loading class to initialize properly JsonPath Cache provider: {}",
-                    Class.forName(JsonPathFunction.class.getName()));
-        } catch (ClassNotFoundException ignored) {
-            LOGGER.trace("Loading class to initialize properly JsonPath Cache provider : fail");
-        }
-    }
 
     private static final String ACCESS_TOKEN_PROPERTY = "access_token";
 
@@ -257,19 +219,8 @@ public class OAuth2AuthenticationResource extends AbstractAuthenticationResource
     }
 
     private Response processUser(final SocialIdentityProviderEntity socialProvider, final HttpServletResponse servletResponse, final String userInfo) {
-        HashMap<String, String> attrs = getUserProfileAttrs(socialProvider.getUserProfileMapping(), userInfo);
-
-        String email = attrs.get(SocialIdentityProviderEntity.UserProfile.EMAIL);
-        if (email == null && socialProvider.isEmailRequired()) {
-            throw new BadRequestException("No public email linked to your account");
-        }
-
-        String userId = null;
-        try {
-            userId = refreshExistingUser(socialProvider, attrs, email);
-        } catch (UserNotFoundException unfe) {
-            userId = createNewExternalUser(socialProvider, userInfo, attrs, email, userId);
-        }
+        UserEntity user = userService.createOrUpdateUserFromSocialIdentityProvider(socialProvider, userInfo);
+        String userId = user.getId();
         
         final Set<RoleEntity> roles =
                 membershipService.getRoles(MembershipReferenceType.PORTAL, singleton("DEFAULT"), userId, RoleScope.PORTAL);
@@ -283,214 +234,10 @@ public class OAuth2AuthenticationResource extends AbstractAuthenticationResource
 
         //set user to Authentication Context
         UserDetails userDetails = new UserDetails(userId, "", authorities);
-        userDetails.setEmail(email);
+        userDetails.setEmail(user.getEmail());
         SecurityContextHolder.getContext().setAuthentication(new UsernamePasswordAuthenticationToken(userDetails, null, userDetails.getAuthorities()));
 
         return connectUser(userId, servletResponse);
-    }
-
-    protected String createNewExternalUser(final SocialIdentityProviderEntity socialProvider, final String userInfo,
-            HashMap<String, String> attrs, String email, String userId) {
-        final NewExternalUserEntity newUser = new NewExternalUserEntity();
-        newUser.setEmail(email);
-        newUser.setSource(socialProvider.getId());
-
-        if (attrs.get(SocialIdentityProviderEntity.UserProfile.ID) != null) {
-            newUser.setSourceId(attrs.get(SocialIdentityProviderEntity.UserProfile.ID));
-        }
-        if (attrs.get(SocialIdentityProviderEntity.UserProfile.LASTNAME) != null) {
-            newUser.setLastname(attrs.get(SocialIdentityProviderEntity.UserProfile.LASTNAME));
-        }
-        if (attrs.get(SocialIdentityProviderEntity.UserProfile.FIRSTNAME) != null) {
-            newUser.setFirstname(attrs.get(SocialIdentityProviderEntity.UserProfile.FIRSTNAME));
-        }
-        if (attrs.get(SocialIdentityProviderEntity.UserProfile.PICTURE) != null) {
-            newUser.setPicture(attrs.get(SocialIdentityProviderEntity.UserProfile.PICTURE));
-        }
-
-        if (socialProvider.getGroupMappings() != null && !socialProvider.getGroupMappings().isEmpty()) {
-            // Can fail if a group in config does not exist in gravitee --> HTTP 500
-            Set<GroupEntity> groupsToAdd = getGroupsToAddUser(userId, socialProvider.getGroupMappings(), userInfo);
-
-            UserEntity createdUser = userService.create(newUser, true);
-            userId = createdUser.getId();
-            addUserToApiAndAppGroupsWithDefaultRole(createdUser.getId(), groupsToAdd);
-        } else {
-            UserEntity createdUser = userService.create(newUser, true);
-            userId = createdUser.getId();
-        }
-
-        if (socialProvider.getRoleMappings() != null && !socialProvider.getRoleMappings().isEmpty()) {
-            Set<RoleEntity> rolesToAdd = getRolesToAddUser(userId, socialProvider.getRoleMappings(), userInfo);
-            addRolesToUser(userId, rolesToAdd);
-        }
-        return userId;
-    }
-
-    protected String refreshExistingUser(final SocialIdentityProviderEntity socialProvider,
-            HashMap<String, String> attrs, String email) {
-        String userId;
-        UserEntity registeredUser = userService.findBySource(socialProvider.getId(), attrs.get(SocialIdentityProviderEntity.UserProfile.ID), false);
-        userId = registeredUser.getId();
-
-        // User refresh
-        UpdateUserEntity user = new UpdateUserEntity();
-
-        if (attrs.get(SocialIdentityProviderEntity.UserProfile.LASTNAME) != null) {
-            user.setLastname(attrs.get(SocialIdentityProviderEntity.UserProfile.LASTNAME));
-        }
-        if (attrs.get(SocialIdentityProviderEntity.UserProfile.FIRSTNAME) != null) {
-            user.setFirstname(attrs.get(SocialIdentityProviderEntity.UserProfile.FIRSTNAME));
-        }
-        if (attrs.get(SocialIdentityProviderEntity.UserProfile.PICTURE) != null) {
-            user.setPicture(attrs.get(SocialIdentityProviderEntity.UserProfile.PICTURE));
-        }
-        user.setEmail(email);
-
-        userService.update(userId, user);
-        return userId;
-    }
-
-    private HashMap<String, String> getUserProfileAttrs(Map<String, String> userProfileMapping, String userInfo) {
-        TemplateEngine templateEngine = TemplateEngine.templateEngine();
-        templateEngine.getTemplateContext().setVariable(TEMPLATE_ENGINE_PROFILE_ATTRIBUTE, userInfo);
-
-        ReadContext userInfoPath = JsonPath.parse(userInfo);
-        HashMap<String, String> map = new HashMap<>(userProfileMapping.size());
-
-        for (Map.Entry<String, String> entry : userProfileMapping.entrySet()) {
-            String field = entry.getKey();
-            String mapping = entry.getValue();
-
-            if (mapping != null) {
-                try {
-                    if (mapping.contains("{#")) {
-                        map.put(field, templateEngine.getValue(mapping, String.class));
-                    } else {
-                        map.put(field, userInfoPath.read(mapping));
-                    }
-                } catch (Exception e) {
-                    LOGGER.error("Using mapping: \"{}\", no fields are located in {}", mapping, userInfo);
-                }
-            }
-        }
-
-        return map;
-    }
-
-    private void addUserToApiAndAppGroupsWithDefaultRole(String userId, Collection<GroupEntity> groupsToAdd) {
-        // Get the default role from system
-        List<RoleEntity> roleEntities = roleService.findDefaultRoleByScopes(RoleScope.API, RoleScope.APPLICATION);
-
-        // Add groups to user
-        for (GroupEntity groupEntity : groupsToAdd) {
-            for (RoleEntity roleEntity : roleEntities) {
-                String defaultRole = roleEntity.getName();
-
-                // If defined, get the override default role at the group level
-                if (groupEntity.getRoles() != null) {
-                    String groupDefaultRole = groupEntity.getRoles().get(io.gravitee.rest.api.model.permissions.RoleScope.valueOf(roleEntity.getScope().name()));
-                    if (groupDefaultRole != null) {
-                        defaultRole = groupDefaultRole;
-                    }
-                }
-
-                membershipService.addOrUpdateMember(
-                        new MembershipService.MembershipReference(MembershipReferenceType.GROUP, groupEntity.getId()),
-                        new MembershipService.MembershipUser(userId, null),
-                        new MembershipService.MembershipRole(mapScope(roleEntity.getScope()), defaultRole));
-            }
-        }
-    }
-
-    private void addRolesToUser(String userId, Collection<RoleEntity> rolesToAdd) {
-        // add roles to user
-        for (RoleEntity roleEntity : rolesToAdd) {
-            membershipService.addOrUpdateMember(
-                    new MembershipService.MembershipReference(
-                            io.gravitee.rest.api.model.permissions.RoleScope.MANAGEMENT == roleEntity.getScope() ?
-                                    MembershipReferenceType.MANAGEMENT : MembershipReferenceType.PORTAL,
-                            MembershipDefaultReferenceId.DEFAULT.name()),
-                    new MembershipService.MembershipUser(userId, null),
-                    new MembershipService.MembershipRole(
-                            RoleScope.valueOf(roleEntity.getScope().name()),
-                            roleEntity.getName()));
-        }
-    }
-
-    private Set<GroupEntity> getGroupsToAddUser(String userId, List<GroupMappingEntity> mappings, String userInfo) {
-        Set<GroupEntity> groupsToAdd = new HashSet<>();
-
-        for (GroupMappingEntity mapping : mappings) {
-            TemplateEngine templateEngine = TemplateEngine.templateEngine();
-            templateEngine.getTemplateContext().setVariable(TEMPLATE_ENGINE_PROFILE_ATTRIBUTE, userInfo);
-
-            boolean match = templateEngine.getValue(mapping.getCondition(), boolean.class);
-
-            trace(userId, match, mapping.getCondition());
-
-            // Get groups
-            if (match) {
-                for (String groupName : mapping.getGroups()) {
-                    try {
-                        groupsToAdd.add(groupService.findById(groupName));
-                    } catch (GroupNotFoundException gnfe) {
-                        LOGGER.error("Unable to create user, missing group in repository : {}", groupName);
-                    }
-                }
-            }
-        }
-        return groupsToAdd;
-    }
-
-    private Set<RoleEntity> getRolesToAddUser(String username, List<RoleMappingEntity> mappings, String userInfo) {
-        Set<RoleEntity> rolesToAdd = new HashSet<>();
-
-        for (RoleMappingEntity mapping : mappings) {
-            TemplateEngine templateEngine = TemplateEngine.templateEngine();
-            templateEngine.getTemplateContext().setVariable(TEMPLATE_ENGINE_PROFILE_ATTRIBUTE, userInfo);
-
-            boolean match = templateEngine.getValue(mapping.getCondition(), boolean.class);
-
-            trace(username, match, mapping.getCondition());
-
-            // Get roles
-            if (match) {
-                addRoleScope(rolesToAdd, mapping.getPortal(), RoleScope.PORTAL);
-                addRoleScope(rolesToAdd, mapping.getManagement(), RoleScope.MANAGEMENT);
-            }
-        }
-        return rolesToAdd;
-    }
-
-    protected void addRoleScope(Set<RoleEntity> rolesToAdd, String mappingRole, RoleScope roleScope) {
-        if (mappingRole != null) {
-            try {
-                RoleEntity roleEntity = roleService.findById(roleScope, mappingRole);
-                rolesToAdd.add(roleEntity);
-            } catch (RoleNotFoundException rnfe) {
-                LOGGER.error("Unable to create user, missing role in repository : {}", mappingRole);
-            }
-        }
-    }
-
-    
-    private void trace(String userId, boolean match, String condition) {
-        if (LOGGER.isDebugEnabled()) {
-            if (match) {
-                LOGGER.debug("the expression {} match on {} user's info ", condition, userId);
-            } else {
-                LOGGER.debug("the expression {} didn't match {} on user's info ", condition, userId);
-            }
-        }
-    }
-
-    private RoleScope mapScope(io.gravitee.rest.api.model.permissions.RoleScope scope) {
-        if (io.gravitee.rest.api.model.permissions.RoleScope.API == scope) {
-            return RoleScope.API;
-        } else {
-            return RoleScope.APPLICATION;
-        }
     }
 
 }
