+++ gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/test/java/io/gravitee/rest/api/portal/rest/resource/auth/OAuth2AuthenticationResourceTest.java
@@ -25,12 +25,9 @@ import static javax.ws.rs.client.Entity.json;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyBoolean;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.ArgumentMatchers.refEq;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
@@ -43,8 +40,6 @@ import java.security.InvalidKeyException;
 import java.security.NoSuchAlgorithmException;
 import java.security.SignatureException;
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -57,7 +52,8 @@ import org.apache.commons.io.IOUtils;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-
+import org.springframework.expression.spel.SpelEvaluationException;
+import org.springframework.expression.spel.SpelMessage;
 import com.auth0.jwt.JWTVerifier;
 import com.auth0.jwt.JWTVerifyException;
 import com.github.tomakehurst.wiremock.client.ResponseDefinitionBuilder;
@@ -65,14 +61,6 @@ import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
 
 import io.gravitee.common.http.HttpStatusCode;
-import io.gravitee.repository.management.model.MembershipDefaultReferenceId;
-import io.gravitee.repository.management.model.MembershipReferenceType;
-import io.gravitee.repository.management.model.RoleScope;
-import io.gravitee.rest.api.model.GroupEntity;
-import io.gravitee.rest.api.model.MemberEntity;
-import io.gravitee.rest.api.model.NewExternalUserEntity;
-import io.gravitee.rest.api.model.RoleEntity;
-import io.gravitee.rest.api.model.UpdateUserEntity;
 import io.gravitee.rest.api.model.UserEntity;
 import io.gravitee.rest.api.model.configuration.identity.GroupMappingEntity;
 import io.gravitee.rest.api.model.configuration.identity.IdentityProviderType;
@@ -81,8 +69,7 @@ import io.gravitee.rest.api.model.configuration.identity.SocialIdentityProviderE
 import io.gravitee.rest.api.portal.rest.model.PayloadInput;
 import io.gravitee.rest.api.portal.rest.model.Token;
 import io.gravitee.rest.api.portal.rest.resource.AbstractResourceTest;
-import io.gravitee.rest.api.service.MembershipService;
-import io.gravitee.rest.api.service.exceptions.UserNotFoundException;
+import io.gravitee.rest.api.service.exceptions.EmailRequiredException;
 
 /**
  * @author Christophe LANNOY (chrislannoy.java at gmail.com)
@@ -209,8 +196,7 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
     }
 
     @Test
-    public void shouldConnectExistingUser() throws Exception {
-
+    public void shouldConnectUser() throws Exception {
         // -- MOCK
         //mock environment
         mockDefaultEnvironment();
@@ -228,16 +214,7 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         userEntity.setSourceId("janedoe@example.com");
         userEntity.setPicture("http://example.com/janedoe/me.jpg");
 
-        when(userService.findBySource(userEntity.getSource(),userEntity.getSourceId(), false)).thenReturn(userEntity);
-
-        //mock DB update user picture , firstname ,lastname
-        UpdateUserEntity user = new UpdateUserEntity();
-        user.setFirstname("Jane");
-        user.setLastname("Doe");
-        user.setEmail("janedoe@example.com");
-        user.setPicture("http://example.com/janedoe/me.jpg");
-
-        when(userService.update(eq(userEntity.getId()), refEq(user))).thenReturn(userEntity);
+        when(userService.createOrUpdateUserFromSocialIdentityProvider(eq(identityProvider), anyString())).thenReturn(userEntity);
 
         //mock DB user connect
         when(userService.connect(userEntity.getId())).thenReturn(userEntity);
@@ -250,9 +227,7 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         Response response = target().request().post(json(payload));
 
         // -- VERIFY
-        verify(userService, times(1)).findBySource(userEntity.getSource(),userEntity.getSourceId(), false);
-        verify(userService, times(0)).create(any());
-        verify(userService, times(1)).update(eq(userEntity.getSourceId()), refEq(user));
+        verify(userService, times(1)).createOrUpdateUserFromSocialIdentityProvider(eq(identityProvider), anyString());
         verify(userService, times(1)).connect(userEntity.getSourceId());
 
         assertEquals(HttpStatusCode.OK_200, response.getStatus());
@@ -299,88 +274,6 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         return payload;
     }
 
-    @Test
-    public void shouldConnectNewUser() throws Exception {
-
-        // -- MOCK
-        //mock environment
-        mockDefaultEnvironment();
-
-        //mock oauth2 exchange authorisation code for access token
-        mockExchangeAuthorizationCodeForAccessToken();
-
-        //mock oauth2 user info call
-        mockUserInfo(okJson(IOUtils.toString(read("/oauth2/json/user_info_response_body.json"), Charset.defaultCharset())));
-
-        //mock DB find user by name
-        when(userService.findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false)).thenThrow(new UserNotFoundException("janedoe@example.com"));
-
-        //mock create user
-        NewExternalUserEntity newExternalUserEntity = mockNewExternalUserEntity();
-        UserEntity createdUser = mockUserEntity();
-        mockUserCreation(newExternalUserEntity, createdUser, true);
-
-        //mock DB update user picture
-        UpdateUserEntity user = mockUpdateUserPicture(createdUser);
-
-        //mock DB user connect
-        when(userService.connect("janedoe@example.com")).thenReturn(createdUser);
-
-
-        // -- CALL
-
-        PayloadInput payload = createPayload("the_client_id","http://localhost/callback","CoDe","StAtE");;
-
-        Response response = target().request().post(json(payload));
-
-        // -- VERIFY
-        verify(userService, times(1)).findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false);
-        verify(userService, times(1)).create(refEq(newExternalUserEntity),eq(true));
-
-//        verify(userService, times(1)).update(eq("janedoe@example.com"), refEq(user));
-        verify(userService, times(1)).connect("janedoe@example.com");
-
-        assertEquals(HttpStatusCode.OK_200, response.getStatus());
-
-        // verify response body
-//        verifyUserInResponseBody(response);
-
-        // verify jwt token
-        verifyJwtToken(response);
-
-    }
-
-    /*
-    private void verifyUserInResponseBody(Response response) {
-        UserEntity responseUser = response.readEntity(UserEntity.class);
-
-        assertEquals(responseUser.getEmail(),"janedoe@example.com");
-        assertEquals(responseUser.getFirstname(),"Jane");
-        assertEquals(responseUser.getLastname(),"Doe");
-        assertEquals(responseUser.getUsername(),"janedoe@example.com");
-        assertEquals(responseUser.getPicture(),"http://example.com/janedoe/me.jpg");
-        assertEquals(responseUser.getSource(),"oauth2");
-    }
-    */
-
-    private UpdateUserEntity mockUpdateUserPicture(UserEntity user) {
-        UpdateUserEntity updateUserEntity = new UpdateUserEntity();
-        updateUserEntity.setPicture("http://example.com/janedoe/me.jpg");
-        updateUserEntity.setFirstname("Jane");
-        updateUserEntity.setLastname("Doe");
-
-        user.setPicture("http://example.com/janedoe/me.jpg");
-        //user.setFirstname("Jane");
-        //user.setLastname("Doe");
-
-        when(userService.update(eq(user.getId()), refEq(updateUserEntity))).thenReturn(user);
-        return updateUserEntity;
-    }
-
-    private void mockUserCreation(NewExternalUserEntity newExternalUserEntity, UserEntity createdUser, boolean addDefaultRole) {
-        when(userService.create(refEq(newExternalUserEntity), eq(addDefaultRole))).thenReturn(createdUser);
-    }
-
     private UserEntity mockUserEntity() {
         UserEntity createdUser = new UserEntity();
         createdUser.setId("janedoe@example.com");
@@ -391,17 +284,6 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         createdUser.setEmail("janedoe@example.com");
         createdUser.setPicture("http://example.com/janedoe/me.jpg");
         return createdUser;
-    }
-
-    private NewExternalUserEntity mockNewExternalUserEntity() {
-        NewExternalUserEntity newExternalUserEntity = new NewExternalUserEntity();
-        newExternalUserEntity.setSource(USER_SOURCE_OAUTH2);
-        newExternalUserEntity.setSourceId("janedoe@example.com");
-        newExternalUserEntity.setLastname("Doe");
-        newExternalUserEntity.setFirstname("Jane");
-        newExternalUserEntity.setEmail("janedoe@example.com");
-        newExternalUserEntity.setPicture("http://example.com/janedoe/me.jpg");
-        return newExternalUserEntity;
     }
 
     @Test
@@ -424,9 +306,7 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         Response response = target().request().post(json(payload));
 
         // -- VERIFY
-        verify(userService, times(0)).findBySource(eq(USER_SOURCE_OAUTH2), anyString(), anyBoolean());
-        verify(userService, times(0)).create(any());
-        verify(userService, times(0)).update(anyString(), any());
+        verify(userService, times(0)).createOrUpdateUserFromSocialIdentityProvider(eq(identityProvider), anyString());
         verify(userService, times(0)).connect(anyString());
 
         assertEquals(HttpStatusCode.UNAUTHORIZED_401, response.getStatus());
@@ -451,6 +331,7 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         //mock oauth2 user info call
         mockUserInfo(okJson(IOUtils.toString(read("/oauth2/json/user_info_response_body.json"), Charset.defaultCharset())));
 
+        when(userService.createOrUpdateUserFromSocialIdentityProvider(refEq(identityProvider),anyString())).thenThrow(new EmailRequiredException(USER_NAME));
         // -- CALL
 
         PayloadInput payload = createPayload("the_client_id","http://localhost/callback","CoDe","StAtE");
@@ -458,9 +339,7 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         Response response = target().request().post(json(payload));
 
         // -- VERIFY
-        verify(userService, times(0)).findBySource(eq(USER_SOURCE_OAUTH2), anyString(), anyBoolean());
-        verify(userService, times(0)).create(any());
-        verify(userService, times(0)).update(anyString(), any());
+        verify(userService, times(1)).createOrUpdateUserFromSocialIdentityProvider(eq(identityProvider), anyString());
         verify(userService, times(0)).connect(anyString());
 
         assertEquals(HttpStatusCode.BAD_REQUEST_400, response.getStatus());
@@ -512,279 +391,11 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
     }
 
     @Test
-    public void shouldConnectNewUserWithGroupsMappingFromUserInfo() throws Exception {
-
-        // -- MOCK
-        //mock environment
-        mockDefaultEnvironment();
-        mockGroupsMapping();
-        mockRolesMapping();
-
-        //mock oauth2 exchange authorisation code for access token
-        mockExchangeAuthorizationCodeForAccessToken();
-
-        //mock oauth2 user info call
-        mockUserInfo(okJson(IOUtils.toString(read("/oauth2/json/user_info_response_body.json"), Charset.defaultCharset())));
-
-        //mock DB find user by name
-        when(userService.findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false)).thenThrow(new UserNotFoundException("janedoe@example.com"));
-
-        //mock create user
-        NewExternalUserEntity newExternalUserEntity = mockNewExternalUserEntity();
-        UserEntity createdUser = mockUserEntity();
-        mockUserCreation(newExternalUserEntity, createdUser, true);
-
-        //mock group search and association
-        when(groupService.findById("Example group")).thenReturn(mockGroupEntity("group_id_1","Example group"));
-        when(groupService.findById("soft user")).thenReturn(mockGroupEntity("group_id_2","soft user"));
-        when(groupService.findById("Others")).thenReturn(mockGroupEntity("group_id_3","Others"));
-        when(groupService.findById("Api consumer")).thenReturn(mockGroupEntity("group_id_4","Api consumer"));
-
-        // mock role search
-        when(roleService.findById(RoleScope.PORTAL, "USER")).thenReturn(mockRoleEntity(io.gravitee.rest.api.model.permissions.RoleScope.PORTAL,"USER"));
-        when(roleService.findById(RoleScope.MANAGEMENT, "ADMIN")).thenReturn(mockRoleEntity(io.gravitee.rest.api.model.permissions.RoleScope.MANAGEMENT,"ADMIN"));
-
-        RoleEntity roleApiUser = mockRoleEntity(io.gravitee.rest.api.model.permissions.RoleScope.API,"USER");
-        RoleEntity roleApplicationAdmin = mockRoleEntity(io.gravitee.rest.api.model.permissions.RoleScope.APPLICATION,"ADMIN");
-
-        when(roleService.findDefaultRoleByScopes(RoleScope.API,RoleScope.APPLICATION)).thenReturn(Arrays.asList(roleApiUser,roleApplicationAdmin));
-
-        when(membershipService.addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_1"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.API, "USER"))).thenReturn(mockMemberEntity());
-
-        when(membershipService.addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_2"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.API, "USER"))).thenReturn(mockMemberEntity());
-
-        when(membershipService.addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_2"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.APPLICATION, "ADMIN"))).thenReturn(mockMemberEntity());
-
-        when(membershipService.addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_4"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.API, "USER"))).thenReturn(mockMemberEntity());
-
-        when(membershipService.addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_4"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.APPLICATION, "ADMIN"))).thenReturn(mockMemberEntity());
-
-        when(membershipService.addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.MANAGEMENT, MembershipDefaultReferenceId.DEFAULT.name()),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.MANAGEMENT, "ADMIN"))).thenReturn(mockMemberEntity());
-
-        when(membershipService.addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.PORTAL, MembershipDefaultReferenceId.DEFAULT.name()),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.PORTAL, "USER"))).thenReturn(mockMemberEntity());
-
-        //mock DB update user picture
-        UpdateUserEntity updateUserEntity = mockUpdateUserPicture(createdUser);
-
-        //mock DB user connect
-        when(userService.connect("janedoe@example.com")).thenReturn(createdUser);
-
-
-        // -- CALL
-
-        PayloadInput payload = createPayload("the_client_id","http://localhost/callback","CoDe","StAtE");
-
-        Response response = target().request().post(json(payload));
-
-        // -- VERIFY
-        verify(userService, times(1)).findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false);
-        verify(userService, times(1)).create(refEq(newExternalUserEntity),eq(true));
-
-//        verify(userService, times(1)).update(eq("janedoe@example.com"), refEq(updateUserEntity));
-        verify(userService, times(1)).connect("janedoe@example.com");
-
-        //verify group creations
-        verify(membershipService, times(1)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_1"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.API, "USER"));
-
-        verify(membershipService, times(1)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_1"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.APPLICATION, "ADMIN"));
-
-        verify(membershipService, times(1)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_2"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.API, "USER"));
-
-        verify(membershipService, times(1)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_2"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.APPLICATION, "ADMIN"));
-
-        verify(membershipService, times(0)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_3"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.API, "USER"));
-
-        verify(membershipService, times(0)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_3"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.APPLICATION, "ADMIN"));
-
-        verify(membershipService, times(1)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_4"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.API, "USER"));
-
-        verify(membershipService, times(1)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.GROUP, "group_id_4"),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.APPLICATION, "ADMIN"));
-
-        verify(membershipService, times(1)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.MANAGEMENT, MembershipDefaultReferenceId.DEFAULT.name()),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.MANAGEMENT, "ADMIN"));
-
-        verify(membershipService, times(1)).addOrUpdateMember(
-                new MembershipService.MembershipReference(MembershipReferenceType.PORTAL, MembershipDefaultReferenceId.DEFAULT.name()),
-                new MembershipService.MembershipUser("janedoe@example.com", null),
-                new MembershipService.MembershipRole(RoleScope.PORTAL, "USER"));
-
-        assertEquals(HttpStatusCode.OK_200, response.getStatus());
-
-        // verify response body
-//        verifyUserInResponseBody(response);
-
-        // verify jwt token
-        verifyJwtToken(response);
-
-    }
-
-    @Test
-    public void shouldConnectNewUserWithNoMatchingGroupsMappingFromUserInfo() throws Exception {
-
-        // -- MOCK
-        //mock environment
-        mockDefaultEnvironment();
-        mockGroupsMapping();
-
-        //mock oauth2 exchange authorisation code for access token
-        mockExchangeAuthorizationCodeForAccessToken();
-
-        //mock oauth2 user info call
-        mockUserInfo(okJson(IOUtils.toString(read("/oauth2/json/user_info_response_body_no_matching.json"), Charset.defaultCharset())));
-
-        //mock DB find user by name
-        when(userService.findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false)).thenThrow(new UserNotFoundException("janedoe@example.com"));
-
-        //mock create user
-        NewExternalUserEntity newExternalUserEntity = mockNewExternalUserEntity();
-        UserEntity createdUser = mockUserEntity();
-        mockUserCreation(newExternalUserEntity, createdUser, true);
-
-        //mock DB update user picture
-        UpdateUserEntity updateUserEntity = mockUpdateUserPicture(createdUser);
-
-        //mock DB user connect
-        when(userService.connect("janedoe@example.com")).thenReturn(createdUser);
-
-
-        // -- CALL
-
-        PayloadInput payload = createPayload("the_client_id","http://localhost/callback","CoDe","StAtE");;
-
-        Response response = target().request().post(json(payload));
-
-        // -- VERIFY
-        verify(userService, times(1)).findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false);
-        verify(userService, times(1)).create(refEq(newExternalUserEntity),eq(true));
-
-//        verify(userService, times(1)).update(eq("janedoe@example.com"), refEq(updateUserEntity));
-        verify(userService, times(1)).connect("janedoe@example.com");
-
-        //verify group creations
-        verify(membershipService, times(0)).addOrUpdateMember(
-                any(MembershipService.MembershipReference.class),
-                any(MembershipService.MembershipUser.class),
-                any(MembershipService.MembershipRole.class));
-
-        assertEquals(HttpStatusCode.OK_200, response.getStatus());
-
-        // verify response body
-//        verifyUserInResponseBody(response);
-
-        // verify jwt token
-        verifyJwtToken(response);
-    }
-
-    @Test
-    public void shouldConnectNewUserWithGroupsMappingFromUserInfoWhenGroupIsNotFound() throws Exception {
-
-        // -- MOCK
-        //mock environment
-        mockDefaultEnvironment();
-        mockGroupsMapping();
-
-        //mock oauth2 exchange authorisation code for access token
-        mockExchangeAuthorizationCodeForAccessToken();
-
-        //mock oauth2 user info call
-        mockUserInfo(okJson(IOUtils.toString(read("/oauth2/json/user_info_response_body.json"), Charset.defaultCharset())));
-
-        //mock DB find user by name
-        when(userService.findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false)).thenThrow(new UserNotFoundException("janedoe@example.com"));
-
-        //mock group search and association
-        when(groupService.findByName("Example group")).thenReturn(Collections.emptyList());
-        when(groupService.findByName("soft user")).thenReturn(Collections.emptyList());
-        when(groupService.findByName("Others")).thenReturn(Collections.emptyList());
-        when(groupService.findByName("Api consumer")).thenReturn(Collections.emptyList());
-
-        NewExternalUserEntity newExternalUserEntity = mockNewExternalUserEntity();
-        UserEntity createdUser = mockUserEntity();
-        mockUserCreation(newExternalUserEntity, createdUser, true);
-
-        //mock DB user connect
-        when(userService.connect(createdUser.getId())).thenReturn(createdUser);
-
-        // -- CALL
-
-        PayloadInput payload = createPayload("the_client_id","http://localhost/callback","CoDe","StAtE");;
-
-        Response response = target().request().post(json(payload));
-
-        // -- VERIFY
-        verify(userService, times(1)).findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false);
-        verify(userService, times(1)).create(any(NewExternalUserEntity.class),anyBoolean());
-
-        verify(userService, times(0)).update(any(String.class), any(UpdateUserEntity.class));
-        verify(userService, times(1)).connect(anyString());
-
-        //verify group creations
-        verify(membershipService, times(0)).addOrUpdateMember(
-                any(MembershipService.MembershipReference.class),
-                any(MembershipService.MembershipUser.class),
-                any(MembershipService.MembershipRole.class));
-
-        assertEquals(HttpStatusCode.OK_200, response.getStatus());
-
-        // verify jwt token
-        verifyJwtToken(response);
-    }
-
-
-    @Test
     public void shouldNotConnectNewUserWhenWrongELGroupsMapping() throws Exception {
 
         // -- MOCK
         //mock environment
         mockDefaultEnvironment();
-        mockWrongELGroupsMapping();
 
         //mock oauth2 exchange authorisation code for access token
         mockExchangeAuthorizationCodeForAccessToken();
@@ -793,11 +404,7 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         mockUserInfo(okJson(IOUtils.toString(read("/oauth2/json/user_info_response_body.json"), Charset.defaultCharset())));
 
         //mock DB find user by name
-        when(userService.findBySource(USER_SOURCE_OAUTH2, "janedoe@example.com",false)).thenThrow(new UserNotFoundException("janedoe@example.com"));
-
-        NewExternalUserEntity newExternalUserEntity = mockNewExternalUserEntity();
-        UserEntity createdUser = mockUserEntity();
-        mockUserCreation(newExternalUserEntity, createdUser, true);
+        when(userService.createOrUpdateUserFromSocialIdentityProvider(eq(identityProvider), anyString())).thenThrow(new SpelEvaluationException(SpelMessage.TYPE_CONVERSION_ERROR, "cannot convert from java.lang.String to boolean"));
 
         // -- CALL
 
@@ -806,96 +413,13 @@ public class OAuth2AuthenticationResourceTest extends AbstractResourceTest {
         Response response = target().request().post(json(payload));
 
         // -- VERIFY
-        verify(userService, times(1)).findBySource(USER_SOURCE_OAUTH2,"janedoe@example.com",false);
-        verify(userService, times(0)).create(any(NewExternalUserEntity.class),anyBoolean());
-
-        verify(userService, times(0)).update(any(String.class), any(UpdateUserEntity.class));
+        verify(userService, times(1)).createOrUpdateUserFromSocialIdentityProvider(eq(identityProvider), anyString());
         verify(userService, times(0)).connect(anyString());
-
-        //verify group creations
-        verify(membershipService, times(0)).addOrUpdateMember(
-                any(MembershipService.MembershipReference.class),
-                any(MembershipService.MembershipUser.class),
-                any(MembershipService.MembershipRole.class));
 
         assertEquals(HttpStatusCode.INTERNAL_SERVER_ERROR_500, response.getStatus());
 
         // verify jwt token
         verifyJwtTokenIsNotPresent(response);
-    }
-
-    private RoleEntity mockRoleEntity(io.gravitee.rest.api.model.permissions.RoleScope scope, String name) {
-        RoleEntity role = new RoleEntity();
-        role.setScope(scope);
-        role.setName(name);
-        return role;
-    }
-
-    private GroupEntity mockGroupEntity( String id, String name) {
-        GroupEntity groupEntity = new GroupEntity();
-        groupEntity.setId(id);
-        groupEntity.setName(name);
-        return groupEntity;
-    }
-
-    private MemberEntity mockMemberEntity() {
-        return mock(MemberEntity.class);
-    }
-
-    private void mockGroupsMapping() {
-
-        GroupMappingEntity condition1 = new GroupMappingEntity();
-        condition1.setCondition("{#jsonPath(#profile, '$.identity_provider_id') == 'idp_5' && #jsonPath(#profile, '$.job_id') != 'API_BREAKER'}");
-        condition1.setGroups(Arrays.asList("Example group", "soft user"));
-        identityProvider.getGroupMappings().add(condition1);
-
-        GroupMappingEntity condition2 = new GroupMappingEntity();
-        condition2.setCondition("{#jsonPath(#profile, '$.identity_provider_id') == 'idp_6'}");
-        condition2.setGroups(Collections.singletonList("Others"));
-        identityProvider.getGroupMappings().add(condition2);
-
-        GroupMappingEntity condition3 = new GroupMappingEntity();
-        condition3.setCondition("{#jsonPath(#profile, '$.job_id') != 'API_BREAKER'}");
-        condition3.setGroups(Collections.singletonList("Api consumer"));
-        identityProvider.getGroupMappings().add(condition3);
-    }
-
-    private void mockRolesMapping() {
-
-        RoleMappingEntity role1 = new RoleMappingEntity();
-        role1.setCondition("{#jsonPath(#profile, '$.identity_provider_id') == 'idp_5' && #jsonPath(#profile, '$.job_id') != 'API_BREAKER'}");
-        role1.setManagement("ADMIN");
-        role1.setPortal("USER");
-        identityProvider.getRoleMappings().add(role1);
-
-        RoleMappingEntity role2 = new RoleMappingEntity();
-        role2.setCondition("{#jsonPath(#profile, '$.identity_provider_id') == 'idp_6'}");
-        role2.setPortal("USER");
-        identityProvider.getRoleMappings().add(role2);
-
-        RoleMappingEntity role3 = new RoleMappingEntity();
-        role3.setCondition("{#jsonPath(#profile, '$.job_id') != 'API_BREAKER'}");
-        role3.setPortal("USER");
-        identityProvider.getRoleMappings().add(role3);
-    }
-
-    private void mockWrongELGroupsMapping() {
-
-        GroupMappingEntity condition1 = new GroupMappingEntity();
-        condition1.setCondition("Some Soup");
-        condition1.setGroups(Arrays.asList("Example group", "soft user"));
-        identityProvider.getGroupMappings().add(condition1);
-
-        GroupMappingEntity condition2 = new GroupMappingEntity();
-        condition2.setCondition("{#jsonPath(#profile, '$.identity_provider_id') == 'idp_6'}");
-        condition2.setGroups(Collections.singletonList("Others"));
-        identityProvider.getGroupMappings().add(condition2);
-
-        GroupMappingEntity condition3 = new GroupMappingEntity();
-        condition3.setCondition("{#jsonPath(#profile, '$.job_id') != 'API_BREAKER'}");
-        condition3.setGroups(Collections.singletonList("Api consumer"));
-        identityProvider.getGroupMappings().add(condition3);
-
     }
 
 }
