+++ gravitee-rest-api-service/src/main/java/io/gravitee/rest/api/service/UserService.java
@@ -19,6 +19,7 @@ import io.gravitee.common.data.domain.Page;
 import io.gravitee.repository.management.api.search.UserCriteria;
 import io.gravitee.rest.api.model.*;
 import io.gravitee.rest.api.model.common.Pageable;
+import io.gravitee.rest.api.model.configuration.identity.SocialIdentityProviderEntity;
 import io.gravitee.rest.api.service.common.JWTHelper.ACTION;
 
 import java.util.List;
@@ -49,4 +50,5 @@ public interface UserService {
     void resetPassword(String id);
     Map<String, Object> getTokenRegistrationParams(UserEntity userEntity, String portalUri, ACTION action);
     UserEntity create(NewExternalUserEntity newExternalUserEntity);
+    UserEntity createOrUpdateUserFromSocialIdentityProvider(SocialIdentityProviderEntity socialProvider, String userInfo);
 }
