+++ gravitee-rest-api-service/src/test/resources/oauth2/json/user_info_response_body.json
@@ -1,0 +1,11 @@
+{
+  "sub":"janedoe@example.com",
+  "name":"Jane Doe",
+  "given_name":"Jane",
+  "family_name":"Doe",
+  "email":"janedoe@example.com",
+  "picture":"http://example.com/janedoe/me.jpg",
+  "identity_provider_id": "idp_5",
+  "job_id": "API_FRIENDLY_USER",
+  "service_id": "585252525"
+}
